Add schema provider to data workspace
