+++ common/data/scripts/generate.ts
@@ -1,0 +1,9 @@
+import { $ } from 'zx'
+import { parseStdout } from '@casimir/zx-helpers'
+
+/**
+ * Generate new sql files from @casimir/data schemas
+ */
+void async function () {
+    console.log('Hello')
+}
