+++ common/data/scripts/watch.ts
@@ -1,0 +1,9 @@
+import { $ } from 'zx'
+import { parseStdout } from '@casimir/zx-helpers'
+
+/**
+ * Watch @casimir/data schemas for changes and generate new sql files
+ */
+void async function () {
+    console.log('Hello')
+}
