+++ contracts/ethereum/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@casimir/ethereum",
     "scripts": {
-        "build": "echo 'Warning: @casimir/ethereum build not specified. Disregard this warning and any listed errors above if @casimir/ethereum is not needed for the current project build.' && exit 0",
+        "build": "echo '@casimir/ethereum build not specified. Disregard this warning and any listed errors above if @casimir/ethereum is not needed for the current project build.' && exit 0",
         "dev": "npx hardhat node",
         "dev:ganache": "npx esno scripts/ganache.ts",
         "deploy": "npx hardhat run scripts/ssv.deploy.ts",
